♻️ refactor: split lecture card to HomeLectureCard and PickLectureCard
