+++ src/entities/lecture/ui/LectureCard/HomeLectureCard/index.ts
@@ -1,0 +1,1 @@
+export { default as HomeLectureCard } from "./HomeLectureCard";
