+++ src/entities/lecture/ui/LectureCard/PickLectureCard/index.ts
@@ -1,0 +1,1 @@
+export { default as PickLectureCard } from "./PickLectureCard";
