+++ src/entities/lecture/ui/LectureCard/RowLectureCard/index.ts
@@ -1,0 +1,1 @@
+export { default as RowLectureCard } from "./RowLectureCard";
