+++ src/entities/lecture/ui/LectureCard/index.ts
@@ -1,1 +1,4 @@
 export { default as LectureCard } from "./LectureCard";
+export * from "./HomeLectureCard";
+export * from "./PickLectureCard";
+export * from "./RowLectureCard";
