+++ src/entities/lecture/ui/LectureList/LectureList.tsx
@@ -8,19 +8,12 @@ interface LectureListProps {
     | LectureInfo[]
     | PickLectureInfo[]
     | HeartsLectureListResDataInfo[];
-  type: "row" | "col";
+  type: "homeLecture" | "pickLecture";
 }
 
 const LectureList = ({ lectureListData, type }: LectureListProps) => {
-  const setClassNameByType = () => {
-    if (type === "col") {
-      return `w-full grid grid-cols-3 gap-6 mobile:grid-cols-1 tablet:grid-cols-2 desktop:grid-cols-3`;
-    }
-    return `w-full grid grid-cols-2 gap-6 mobile:grid-cols-1 tablet:grid-cols-1 desktop:grid-cols-2 pb-[265px]`;
-  };
-
   return (
-    <div className={setClassNameByType()}>
+    <div className="w-full grid grid-cols-3 gap-6 mobile:grid-cols-1 tablet:grid-cols-2 desktop:grid-cols-3">
       {lectureListData.map((lectureData) => {
         return (
           <LectureCard
